get rid of roles
Unfortunately, roles were a regrettably overengineered abomination.
Noone except for me (and me neither) could understand
how they work, they required a lot of boilerplate - shortly put,
they didn't pull their weight. They needed to go.

There were two strategies that I used during this refactoring:

1) Instead of attaching a role to a tree and then using this role in
a subsequent extractor, I would create an instance of a dedicated class
and then have a dedicated extractor for it. Examples: TermParamDef, TypeParamDef, etc.

2) Otherwise, use a different way to mark trees (see the `patterns` map
at the beginning of LogicalTrees.scala).
